Convert eregs into a Django management command
We'll want access to much of Django's settings, particularly around database
setup going forward. As such, we should execute our parser commands within
Django's context. The easiest way to implement that is by converting the
commands into Django management commands.

Keeps eregs.py script for backwards compatibility, though we'll probably want
to remove it.

You can now execute a parsing command in one of four ways:

* eregs pipeline 27 646 output
* python eregs.py pipeline 27 646 output
* python manage.py eregs pipeline 27 646 output
* python manage.py async_eregs pipeline 27 646 output
