+++ requirements.txt
@@ -2,6 +2,8 @@ boto3==1.3.1
 cached-property==1.3.0
 click==6.6
 coloredlogs==5.0
+Django==1.9.7
+django-click==1.2.0
 GitPython==2.0.2
 inflection==0.3.1
 ipdb==0.10.0
