+++ requirements_web.txt
@@ -1,4 +1,3 @@
-Django==1.9.7
 redis==2.10.5
 rq==0.6.0
 django_rq==0.9.1
