+++ setup.py
@@ -23,5 +23,6 @@ setup(
         "requests",
         "requests-cache"
     ],
-    entry_points={"console_scripts": ["eregs=eregs:cli"]}
+    entry_points={"console_scripts": [
+        "eregs=regparser.web.management.commands.eregs:cli.main"]}
 )
